+++ AmigoPJT/AmigoPJT.vcxproj
@@ -145,7 +145,7 @@
     <ClCompile Include="command.cpp" />
     <ClCompile Include="InputStream.cpp" />
     <ClCompile Include="main.cpp" />
-    <ClCompile Include="search1.cpp" />
+    <ClCompile Include="searchByBirthday.cpp" />
     <ClCompile Include="searchByName.cpp" />
     <ClCompile Include="searchByPhoneNumber.cpp" />
     <ClCompile Include="searchByCerti.cpp" />
@@ -159,7 +159,7 @@
     <ClInclude Include="command.h" />
     <ClInclude Include="common.h" />
     <ClInclude Include="inputStream.h" />
-    <ClInclude Include="search1.h" />
+    <ClInclude Include="searchByBirthday.h" />
     <ClInclude Include="searchByName.h" />
     <ClInclude Include="searchByPhoneNumber.h" />
     <ClInclude Include="searchEtc.h" />
