+++ AmigoPJT/AmigoPJT.vcxproj.filters
@@ -24,7 +24,7 @@
     <ClCompile Include="command\mod.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
-    <ClCompile Include="search1.cpp">
+    <ClCompile Include="searchByBirthday.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
     <ClCompile Include="searchByName.cpp">
@@ -57,6 +57,9 @@
     <ClCompile Include="InputStream.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
+    <ClCompile Include="client.h">
+      <Filter>헤더 파일</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <Text Include="input_20_20.txt">
@@ -70,9 +73,6 @@
     <ClInclude Include="cmd_parser.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
-    <ClInclude Include="client.h">
-      <Filter>헤더 파일</Filter>
-    </ClInclude>
     <ClInclude Include="command.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
@@ -82,7 +82,7 @@
     <ClInclude Include="command\mod.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
-    <ClInclude Include="search1.h">
+    <ClInclude Include="searchByBirthday.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
     <ClInclude Include="searchByName.h">
