+++ AmigoPJT/Command.h
@@ -2,7 +2,7 @@
 #include <iostream>
 #include <unordered_map>
 #include "common.h"
-#include "search1.h"
+#include "searchByBirthday.h"
 #include "command/mod.h"
 #include "searchByName.h"
 #include "searchByPhoneNumber.h"
@@ -21,5 +21,5 @@ int Mod(const vector<unsigned int>& founds, string column, string value);
 string GenerateCommandRecord(const std::string& command, const bool& detail_print, const vector<unsigned int>& targets);
 string GenerateDetailRecord(const std::string& command, const vector<unsigned int>& result);
 
-void CommandRun(vector<Command> commands);
+vector<string> CommandRun(vector<Command> commands);
 string Del(unordered_map<unsigned int, Employee2>& employee, vector<unsigned int> deleteList);
